+++ lib/src/qr.dart
@@ -11,7 +11,6 @@ import 'routes/qroute_children.dart';
 import 'routes/qroute_internal.dart';
 import 'types/qhistory.dart';
 import 'types/qobserver.dart';
-import 'types/route_mock.dart';
 
 class QRContext {
   static const rootRouterName = 'Root';